Extract event handling functions into module
